little fixes
